backend/requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-environ==0.8.1
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-nested-routers==0.93.4
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-secret-manager==2.7.2
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.41.1
grpcio-status==1.41.1
idna==3.3
libcst==0.3.21
mypy-extensions==0.4.3
packaging==21.2
proto-plus==1.19.7
protobuf==3.19.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.3.0
pyparsing==2.4.7
pytz==2021.3
PyYAML==6.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
typing-inspect==0.7.1
urllib3==1.26.7
